feat: Control behaviour from the command line
3 CLI arguments were added in this commit to allow better control of the
code. These replacing hard-coded settings.

The verbosity is a useful switch in general. Smaller projects might not
even need it, but if the project is a sufficiently big, the user might
want to have some idea what is happening under the hood.

The library-path/CLANG_LIBRARY_PATH is added so that the user can easily
configure where the clang libraries can be found. The env variable makes
it easier to have a persistent setting in the system. I use direnv where
I can define the variable.

Finally, the max-sources argument is there to allow faster feedback for
code changes and debugging.