fix: Build fully qualified function names
Inferring the qualification from the AST traversal is not reliable. clang
however, offers a way to walk backward to the "semantic_parent" tree,
which produces almost the correct parentage.

There seems to be a ID mismatch with the TRANSLATION_UNIT kind which is
now reported to be 350 and "unknown" to libclang python module. I'm
pretty sure that this is a problem with some version mismatch but I
still need to prove it.